// routes/logic/classes/interfaces/ImportTypesInterface.ts
/** reviewed by Ronny Brandt - 24.07.2021
 * beautiful file, beautiful code, nothing messy here */

export interface ImportTypesInterface {
  importTypeId: number;
  importTypeName: string;
  importTypeActive: number;
}

export interface ImportTypesDataInterface {
  ID: number;
  IMPORT_TYPE_NAME: string;
  ACTIVE: number;
}